+++ SFSymbolsPreview/iOS/Views/CategoriesList.swift
@@ -31,7 +31,8 @@ struct CategoriesList: View {
             }
         }
         .navigationTitle("Categories")
-        .searchable(text: $searchText)        
+        .searchable(text: $searchText, placement: .sidebar)
+        .animation(.default, value: searchText)
     }
 }