+++ SFSymbolsPreview/iOS/Views/InitialDetailsView.swift
@@ -35,8 +35,9 @@ struct InitialDetailsView: View{
             }
             .textCase(nil)
         }
-        .searchable(text: $searchText)
+        .searchable(text: $searchText, placement: .sidebar)
         .navigationTitle("Symbols")
+        .animation(.default, value: searchText)
     }
 }