+++ Sources/SFSymbols/SFSymbol+StaticVariables/SFSymbol+StaticVariables14P2.swift
@@ -1,0 +1,484 @@
+import Foundation
+
+@available(iOS 14.2, macOS 11.0, tvOS 14.2, watchOS 7.1, *)
+public extension SFSymbol {
+    static let aqiHigh = SFSymbol(title: "aqi.high",
+                                        categories: [.multicolor, .weather],
+                                        searchTerms: ["weather"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let aqiLow = SFSymbol(title: "aqi.low",
+                                        categories: [.weather],
+                                        searchTerms: ["weather"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let aqiMedium = SFSymbol(title: "aqi.medium",
+                                        categories: [.weather],
+                                        searchTerms: ["weather"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let brazilianrealsignCircle = SFSymbol(title: "brazilianrealsign.circle",
+                                        categories: [.commerce, .indices],
+                                        searchTerms: ["currencies", "currency"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let brazilianrealsignCircleFill = SFSymbol(title: "brazilianrealsign.circle.fill",
+                                        categories: [.commerce, .indices, .multicolor],
+                                        searchTerms: ["currencies", "currency"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let brazilianrealsignSquare = SFSymbol(title: "brazilianrealsign.square",
+                                        categories: [.commerce, .indices],
+                                        searchTerms: ["currencies", "currency"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let brazilianrealsignSquareFill = SFSymbol(title: "brazilianrealsign.square.fill",
+                                        categories: [.commerce, .indices, .multicolor],
+                                        searchTerms: ["currencies", "currency"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let cartCircle = SFSymbol(title: "cart.circle",
+                                        categories: [.commerce],
+                                        searchTerms: ["shopping.cart.circle"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let cartCircleFill = SFSymbol(title: "cart.circle.fill",
+                                        categories: [.commerce, .multicolor],
+                                        searchTerms: ["shopping.cart.circle.fill"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let character = SFSymbol(title: "character",
+                                        categories: [.textformatting],
+                                        searchTerms: ["a"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterAr = SFSymbol(title: "character.ar",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosed = SFSymbol(title: "character.book.closed",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedAr = SFSymbol(title: "character.book.closed.ar",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedFill = SFSymbol(title: "character.book.closed.fill",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedFillAr = SFSymbol(title: "character.book.closed.fill.ar",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedFillHe = SFSymbol(title: "character.book.closed.fill.he",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedFillHi = SFSymbol(title: "character.book.closed.fill.hi",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedFillJa = SFSymbol(title: "character.book.closed.fill.ja",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedFillKo = SFSymbol(title: "character.book.closed.fill.ko",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedFillTh = SFSymbol(title: "character.book.closed.fill.th",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedHe = SFSymbol(title: "character.book.closed.he",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedHi = SFSymbol(title: "character.book.closed.hi",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedJa = SFSymbol(title: "character.book.closed.ja",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedKo = SFSymbol(title: "character.book.closed.ko",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterBookClosedTh = SFSymbol(title: "character.book.closed.th",
+                                        categories: [.objectsandtools],
+                                        searchTerms: ["dictionary"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterHe = SFSymbol(title: "character.he",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterHi = SFSymbol(title: "character.hi",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterJa = SFSymbol(title: "character.ja",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterKo = SFSymbol(title: "character.ko",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let characterTh = SFSymbol(title: "character.th",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let clockArrow2Circlepath = SFSymbol(title: "clock.arrow.2.circlepath",
+                                        categories: [.time],
+                                        searchTerms: ["sync", "time", "time machine", "timemachine"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let directcurrent = SFSymbol(title: "directcurrent",
+                                        categories: nil,
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let docTextBelowEcgFillRtl = SFSymbol(title: "doc.text.below.ecg.fill.rtl",
+                                        categories: nil,
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let docTextBelowEcgRtl = SFSymbol(title: "doc.text.below.ecg.rtl",
+                                        categories: nil,
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let exclamationmarkApplewatch = SFSymbol(title: "exclamationmark.applewatch",
+                                        categories: [.devices],
+                                        searchTerms: ["warning"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let infinityCircle = SFSymbol(title: "infinity.circle",
+                                        categories: [.media],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let infinityCircleFill = SFSymbol(title: "infinity.circle.fill",
+                                        categories: [.media, .multicolor],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let ipadBadgePlay = SFSymbol(title: "ipad.badge.play",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let ipadHomebuttonBadgePlay = SFSymbol(title: "ipad.homebutton.badge.play",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let ipadHomebuttonLandscapeBadgePlay = SFSymbol(title: "ipad.homebutton.landscape.badge.play",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let ipadLandscapeBadgePlay = SFSymbol(title: "ipad.landscape.badge.play",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let iphoneBadgePlay = SFSymbol(title: "iphone.badge.play",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let iphoneHomebuttonBadgePlay = SFSymbol(title: "iphone.homebutton.badge.play",
+                                        categories: [.devices],
+                                        searchTerms: ["keynote.remote"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let iphoneHomebuttonLandscape = SFSymbol(title: "iphone.homebutton.landscape",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let iphoneLandscape = SFSymbol(title: "iphone.landscape",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let ipodtouchLandscape = SFSymbol(title: "ipodtouch.landscape",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let lockApplewatch = SFSymbol(title: "lock.applewatch",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let photoTv = SFSymbol(title: "photo.tv",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let playTv = SFSymbol(title: "play.tv",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let playTvFill = SFSymbol(title: "play.tv.fill",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let rectangleBadgePersonCrop = SFSymbol(title: "rectangle.badge.person.crop",
+                                        categories: [.human],
+                                        searchTerms: ["people"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let rectangleFillBadgePersonCrop = SFSymbol(title: "rectangle.fill.badge.person.crop",
+                                        categories: [.human],
+                                        searchTerms: ["people"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let rectangleTopthirdInset = SFSymbol(title: "rectangle.topthird.inset",
+                                        categories: nil,
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let repeat1Circle = SFSymbol(title: "repeat.1.circle",
+                                        categories: [.media],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let repeat1CircleFill = SFSymbol(title: "repeat.1.circle.fill",
+                                        categories: [.media, .multicolor],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let repeatCircle = SFSymbol(title: "repeat.circle",
+                                        categories: [.media],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let repeatCircleFill = SFSymbol(title: "repeat.circle.fill",
+                                        categories: [.media, .multicolor],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let shieldCheckerboard = SFSymbol(title: "shield.checkerboard",
+                                        categories: nil,
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let shuffleCircle = SFSymbol(title: "shuffle.circle",
+                                        categories: [.media],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let shuffleCircleFill = SFSymbol(title: "shuffle.circle.fill",
+                                        categories: [.media, .multicolor],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textBelowPhotoFillRtl = SFSymbol(title: "text.below.photo.fill.rtl",
+                                        categories: nil,
+                                        searchTerms: ["environment", "image", "mountain", "place", "sun"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textBelowPhotoRtl = SFSymbol(title: "text.below.photo.rtl",
+                                        categories: nil,
+                                        searchTerms: ["environment", "image", "mountain", "place", "sun"],
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeAr = SFSymbol(title: "textformat.size.ar",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeHe = SFSymbol(title: "textformat.size.he",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeHi = SFSymbol(title: "textformat.size.hi",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeJa = SFSymbol(title: "textformat.size.ja",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeKo = SFSymbol(title: "textformat.size.ko",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeLarger = SFSymbol(title: "textformat.size.larger",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeLargerAr = SFSymbol(title: "textformat.size.larger.ar",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeLargerHe = SFSymbol(title: "textformat.size.larger.he",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeLargerHi = SFSymbol(title: "textformat.size.larger.hi",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeLargerJa = SFSymbol(title: "textformat.size.larger.ja",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeLargerKo = SFSymbol(title: "textformat.size.larger.ko",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeLargerTh = SFSymbol(title: "textformat.size.larger.th",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeSmaller = SFSymbol(title: "textformat.size.smaller",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeSmallerAr = SFSymbol(title: "textformat.size.smaller.ar",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeSmallerHe = SFSymbol(title: "textformat.size.smaller.he",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeSmallerHi = SFSymbol(title: "textformat.size.smaller.hi",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeSmallerJa = SFSymbol(title: "textformat.size.smaller.ja",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeSmallerKo = SFSymbol(title: "textformat.size.smaller.ko",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeSmallerTh = SFSymbol(title: "textformat.size.smaller.th",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSizeTh = SFSymbol(title: "textformat.size.th",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSubscriptAr = SFSymbol(title: "textformat.subscript.ar",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSubscriptHe = SFSymbol(title: "textformat.subscript.he",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSubscriptHi = SFSymbol(title: "textformat.subscript.hi",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSubscriptJa = SFSymbol(title: "textformat.subscript.ja",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSubscriptKo = SFSymbol(title: "textformat.subscript.ko",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSubscriptTh = SFSymbol(title: "textformat.subscript.th",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSuperscriptAr = SFSymbol(title: "textformat.superscript.ar",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSuperscriptHe = SFSymbol(title: "textformat.superscript.he",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSuperscriptHi = SFSymbol(title: "textformat.superscript.hi",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSuperscriptJa = SFSymbol(title: "textformat.superscript.ja",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSuperscriptKo = SFSymbol(title: "textformat.superscript.ko",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let textformatSuperscriptTh = SFSymbol(title: "textformat.superscript.th",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let torus = SFSymbol(title: "torus",
+                                        categories: nil,
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+
+    static let tvAndMediabox = SFSymbol(title: "tv.and.mediabox",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 14.2, macOS: 11.0, tvOS: 14.2, watchOS: 7.1))
+}