+++ Sources/SFSymbols/SFSymbol+StaticVariables/SFSymbol+StaticVariables15P1.swift
@@ -1,0 +1,69 @@
+import Foundation
+
+@available(iOS 15.1, macOS 12.0, tvOS 15.1, watchOS 8.1, *)
+public extension SFSymbol {
+    static let boltRingClosed = SFSymbol(title: "bolt.ring.closed",
+                                        categories: nil,
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let platterFilledBottomAndArrowDownIphone = SFSymbol(title: "platter.filled.bottom.and.arrow.down.iphone",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let platterFilledBottomIphone = SFSymbol(title: "platter.filled.bottom.iphone",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let platterFilledTopAndArrowUpIphone = SFSymbol(title: "platter.filled.top.and.arrow.up.iphone",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let platterFilledTopIphone = SFSymbol(title: "platter.filled.top.iphone",
+                                        categories: [.devices],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let square3Layers3DDownBackward = SFSymbol(title: "square.3.layers.3d.down.backward",
+                                        categories: nil,
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let square3Layers3DDownForward = SFSymbol(title: "square.3.layers.3d.down.forward",
+                                        categories: nil,
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let square3Layers3DDownLeft = SFSymbol(title: "square.3.layers.3d.down.left",
+                                        categories: nil,
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let square3Layers3DDownRight = SFSymbol(title: "square.3.layers.3d.down.right",
+                                        categories: nil,
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let textJustifyLeading = SFSymbol(title: "text.justify.leading",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let textJustifyLeft = SFSymbol(title: "text.justify.left",
+                                        categories: [.textformatting],
+                                        searchTerms: ["justify.left"],
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let textJustifyRight = SFSymbol(title: "text.justify.right",
+                                        categories: [.textformatting],
+                                        searchTerms: ["justify.right"],
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+
+    static let textJustifyTrailing = SFSymbol(title: "text.justify.trailing",
+                                        categories: [.textformatting],
+                                        searchTerms: nil,
+                                        releaseInfo: ReleaseInfo(iOS: 15.1, macOS: 12.0, tvOS: 15.1, watchOS: 8.1))
+}